+++ src/commonMain/kotlin/com/gw2tb/manager/discoverer/AddOnDiscoverer.kt
@@ -18,6 +18,7 @@ package com.gw2tb.manager.discoverer
 
 import com.gw2tb.manager.model.local.LocalAddOn
 import java.nio.file.Path
+import kotlin.io.path.isSameFileAs
 
 /**
  * An add-on discoverer is responsible for the discovery and inspection of local add-ons.
@@ -30,10 +31,24 @@ interface AddOnDiscoverer {
     /**
      * Returns a list of local add-ons found in the given game directory.
      *
-     * @param gameDirectory the game directory to search for add-ons
+     * @param gameDirectory     the game directory to search for add-ons
+     * @param discoveredAddOns  the list of add-ons that has already been discovered (typically, by other discoverers
+     *                          earlier in the chain)
      *
      * @return  a list of local add-ons found in the given game directory
      */
-    fun getAddOns(gameDirectory: Path): List<LocalAddOn>
+    fun getAddOns(gameDirectory: Path, discoveredAddOns: List<LocalAddOn>): List<LocalAddOn>
+
+    /**
+     * Returns a filtering function that checks whether a given path refers to the same file as a previously discovered
+     * add-on.
+     *
+     * @param discoveredAddOns  the previously discovered add-ons to consider
+     *
+     * @return  the filter function
+     */
+    fun isDistinctFrom(discoveredAddOns: List<LocalAddOn>): (Path) -> Boolean = { path ->
+        discoveredAddOns.none { localAddOn -> localAddOn.path.isSameFileAs(path) }
+    }
 
 }